:sparkls: add exclude files

+++ deploy.sh
@@ -29,7 +29,7 @@ create_dotfiles_symlink() {
       continue
     fi
 
-    if [[ ! -e $dotfile ]]; then
+    if [[ ! -e $dotfile ]] && [[ $dotfile == 'README.md' ]] && [[ $dotfile == 'deploy.sh' ]]; then
       echo "Make symlink $src"
       ln -s $src $dotfile
     fi
